Fix count_adds

+++ src/include/libpmemobj.h
@@ -812,7 +812,7 @@ void pmemobj_tx_commit(void);
 void pmemobj_tx_commit_group(PMEMobjpool *pop, jmp_buf env);
 void pmemobj_tx_end_group(void);
 
-int count_adds(void);
+uint64_t count_adds(void);
 
 /*
  * Cleanups current transaction. Must always be called after pmemobj_tx_begin,
